feat: move logout to main layout
